Refactor extension architecture into modular src layout
