+++ src/options/index.html
@@ -84,6 +84,6 @@
         <div id="status" class="status" style="display: none;"></div>
     </div>
 
-    <script src="options.js"></script>
+    <script src="index.js"></script>
 </body>
 </html>
