+++ src/ui/index.html
@@ -123,6 +123,6 @@
         • 執行過程中請勿關閉相關分頁
     </div>
 
-    <script src="popup.js"></script>
+    <script src="index.js"></script>
 </body>
 </html>
